backend/requirements.txt
flask
requests
cryptography
python-dotenv
